requirements.txt
itchat>=1.3.10
echarts-python==0.1.3
jieba==0.39
matplotlib==2.1.1
numpy==1.13.3